+++ tests/Ritten.Tests/Changelogs/CheckChangelogLinksTests.cs
@@ -24,7 +24,7 @@ public class CheckChangelogLinksTests
         .Services.BuildServiceProvider()
         .GetRequiredService<IChangelog>();
 
-    private readonly IBuildReport _report = Substitute.For<IBuildReport>();
+    private readonly IWorkflowReport _report = Substitute.For<IWorkflowReport>();
     private readonly ReportSection _changelogSection = new("Changelog");
     private readonly ChangelogOptions _options = TestOptions.Changelog();