+++ tests/Ritten.Tests/Changelogs/ReadChangelogTests.cs
@@ -22,7 +22,7 @@ public class ReadChangelogTests
         .GetRequiredService<IChangelog>();
 
     private readonly IFileSystem _fileSystem = Substitute.For<IFileSystem>();
-    private readonly IBuildReport _report = Substitute.For<IBuildReport>();
+    private readonly IWorkflowReport _report = Substitute.For<IWorkflowReport>();
     private readonly ReportSection _changelogSection = new("Changelog");
     private readonly ChangelogOptions _options = TestOptions.Changelog();