+++ tests/Ritten.Tests/Extensions/WorkflowBuilderExtensionsTests.cs
@@ -47,7 +47,7 @@ public class WorkflowBuilderExtensionsTests
         services.Count(d => d.ServiceType == typeof(INuGet)).ShouldBe(1);
         services.Count(d => d.ServiceType == typeof(IGitHubClient)).ShouldBe(1);
         services.Count(d => d.ServiceType == typeof(IReleaseService)).ShouldBe(1);
-        services.Count(d => d.ServiceType == typeof(IBuildReport)).ShouldBe(1);
+        services.Count(d => d.ServiceType == typeof(IWorkflowReport)).ShouldBe(1);
     }
 
     [Fact]