+++ CHANGELOG.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [Unreleased]
+
+### Changed
+
+- **IBuildReport is now IWorkflowReport.** It more accurately reflects that the report covers one workflow execution.
+
 ## [0.3.0] - 2026-08-19
 
 ### Changed